Correct mistake where `duo` and `yubikey_token` are valid key values although they may only be available by feature flag.
